+++ pkg/ring/ring_test.go
@@ -333,7 +333,7 @@ func TestRing_ShuffleShard(t *testing.T) {
 
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
-			// Init the ring description.
+			// Init the ring.
 			ringDesc := &Desc{Ingesters: testData.ringInstances}
 			for id, instance := range ringDesc.Ingesters {
 				instance.Timestamp = time.Now().Unix()
@@ -512,8 +512,8 @@ func TestRing_ShuffleShard_Shuffling(t *testing.T) {
 	maxCombinations := int(math.Pow(float64(numTenants), 2)) - numTenants
 	for numMatching, probability := range theoreticalMatchings {
 		// We allow a max deviance of 10% compared to the theoretical probability,
-		// clamping it between 1% and 0.1% boundaries.
-		maxDeviance := math.Min(1, math.Max(0.1, probability*0.1))
+		// clamping it between 1% and 0.2% boundaries.
+		maxDeviance := math.Min(1, math.Max(0.2, probability*0.1))
 
 		actual := (float64(distribution[numMatching]) / float64(maxCombinations)) * 100
 		assert.InDelta(t, probability, actual, maxDeviance, "numMatching: %d", numMatching)
@@ -606,6 +606,156 @@ func TestRing_ShuffleShard_Consistency(t *testing.T) {
 				assert.LessOrEqual(t, len(removed), 1)
 			}
 		})
+	}
+}
+
+func TestRing_ShuffleShard_ConsistencyOnShardSizeChanged(t *testing.T) {
+	// Create 30 instances in 3 zones.
+	ringInstances := map[string]IngesterDesc{}
+	for i := 0; i < 30; i++ {
+		name, desc := generateRingInstance(i, i%3)
+		ringInstances[name] = desc
+	}
+
+	// Init the ring.
+	ringDesc := &Desc{Ingesters: ringInstances}
+	ring := Ring{
+		cfg: Config{
+			HeartbeatTimeout:     time.Hour,
+			ZoneAwarenessEnabled: true,
+		},
+		ringDesc:         ringDesc,
+		ringTokens:       ringDesc.getTokens(),
+		ringTokensByZone: ringDesc.getTokensByZone(),
+		ringZones:        getZones(ringDesc.getTokensByZone()),
+		strategy:         &DefaultReplicationStrategy{},
+	}
+
+	// Get the replication set with shard size = 3.
+	firstShard := ring.ShuffleShard("tenant-id", 3)
+	assert.Equal(t, 3, firstShard.IngesterCount())
+
+	firstSet, err := firstShard.GetAll(Read)
+	require.NoError(t, err)
+
+	// Increase shard size to 6.
+	secondShard := ring.ShuffleShard("tenant-id", 6)
+	assert.Equal(t, 6, secondShard.IngesterCount())
+
+	secondSet, err := secondShard.GetAll(Read)
+	require.NoError(t, err)
+
+	for _, firstInstance := range firstSet.Ingesters {
+		assert.True(t, secondSet.Includes(firstInstance.Addr), "new replication set is expected to include previous instance %s", firstInstance.Addr)
+	}
+
+	// Increase shard size to 9.
+	thirdShard := ring.ShuffleShard("tenant-id", 9)
+	assert.Equal(t, 9, thirdShard.IngesterCount())
+
+	thirdSet, err := thirdShard.GetAll(Read)
+	require.NoError(t, err)
+
+	for _, secondInstance := range secondSet.Ingesters {
+		assert.True(t, thirdSet.Includes(secondInstance.Addr), "new replication set is expected to include previous instance %s", secondInstance.Addr)
+	}
+
+	// Decrease shard size to 6.
+	fourthShard := ring.ShuffleShard("tenant-id", 6)
+	assert.Equal(t, 6, fourthShard.IngesterCount())
+
+	fourthSet, err := fourthShard.GetAll(Read)
+	require.NoError(t, err)
+
+	// We expect to have the same exact instances we had when the shard size was 6.
+	for _, secondInstance := range secondSet.Ingesters {
+		assert.True(t, fourthSet.Includes(secondInstance.Addr), "new replication set is expected to include previous instance %s", secondInstance.Addr)
+	}
+
+	// Decrease shard size to 3.
+	fifthShard := ring.ShuffleShard("tenant-id", 3)
+	assert.Equal(t, 3, fifthShard.IngesterCount())
+
+	fifthSet, err := fifthShard.GetAll(Read)
+	require.NoError(t, err)
+
+	// We expect to have the same exact instances we had when the shard size was 3.
+	for _, firstInstance := range firstSet.Ingesters {
+		assert.True(t, fifthSet.Includes(firstInstance.Addr), "new replication set is expected to include previous instance %s", firstInstance.Addr)
+	}
+}
+
+func TestRing_ShuffleShard_ConsistencyOnZonesChanged(t *testing.T) {
+	// Create 20 instances in 2 zones.
+	ringInstances := map[string]IngesterDesc{}
+	for i := 0; i < 20; i++ {
+		name, desc := generateRingInstance(i, i%2)
+		ringInstances[name] = desc
+	}
+
+	// Init the ring.
+	ringDesc := &Desc{Ingesters: ringInstances}
+	ring := Ring{
+		cfg: Config{
+			HeartbeatTimeout:     time.Hour,
+			ZoneAwarenessEnabled: true,
+		},
+		ringDesc:         ringDesc,
+		ringTokens:       ringDesc.getTokens(),
+		ringTokensByZone: ringDesc.getTokensByZone(),
+		ringZones:        getZones(ringDesc.getTokensByZone()),
+		strategy:         &DefaultReplicationStrategy{},
+	}
+
+	// Get the replication set with shard size = 2.
+	firstShard := ring.ShuffleShard("tenant-id", 2)
+	assert.Equal(t, 2, firstShard.IngesterCount())
+
+	firstSet, err := firstShard.GetAll(Read)
+	require.NoError(t, err)
+
+	// Increase shard size to 4.
+	secondShard := ring.ShuffleShard("tenant-id", 4)
+	assert.Equal(t, 4, secondShard.IngesterCount())
+
+	secondSet, err := secondShard.GetAll(Read)
+	require.NoError(t, err)
+
+	for _, firstInstance := range firstSet.Ingesters {
+		assert.True(t, secondSet.Includes(firstInstance.Addr), "new replication set is expected to include previous instance %s", firstInstance.Addr)
+	}
+
+	// Scale up cluster, adding 10 instances in 1 new zone.
+	for i := 20; i < 30; i++ {
+		name, desc := generateRingInstance(i, 2)
+		ringInstances[name] = desc
+	}
+
+	ring.ringDesc.Ingesters = ringInstances
+	ring.ringTokens = ringDesc.getTokens()
+	ring.ringTokensByZone = ringDesc.getTokensByZone()
+	ring.ringZones = getZones(ringDesc.getTokensByZone())
+
+	// Increase shard size to 6.
+	thirdShard := ring.ShuffleShard("tenant-id", 6)
+	assert.Equal(t, 6, thirdShard.IngesterCount())
+
+	thirdSet, err := thirdShard.GetAll(Read)
+	require.NoError(t, err)
+
+	for _, secondInstance := range secondSet.Ingesters {
+		assert.True(t, thirdSet.Includes(secondInstance.Addr), "new replication set is expected to include previous instance %s", secondInstance.Addr)
+	}
+
+	// Increase shard size to 9.
+	fourthShard := ring.ShuffleShard("tenant-id", 9)
+	assert.Equal(t, 9, fourthShard.IngesterCount())
+
+	fourthSet, err := fourthShard.GetAll(Read)
+	require.NoError(t, err)
+
+	for _, thirdInstance := range thirdSet.Ingesters {
+		assert.True(t, fourthSet.Includes(thirdInstance.Addr), "new replication set is expected to include previous instance %s", thirdInstance.Addr)
 	}
 }